+++ Scripts/5_Mission/PaintZ/PaintZ_MissionServer.c
@@ -3,12 +3,15 @@ modded class MissionServer
     override void OnInit()
     {
         super.OnInit();
+        PaintZ_RuntimeDocs.Refresh();
+        PaintZ_PatternScaling.StartServer();
         PaintZ_ItemPolicy.StartServer();
     }
 
     override void OnMissionFinish()
     {
         PaintZ_ItemPolicy.StopServer();
+        PaintZ_PatternScaling.StopServer();
         super.OnMissionFinish();
     }